+++ opticspy/__init__.py
@@ -9,7 +9,7 @@
  *  Copyright (c) 2014-2015 Xing fan
 """
 from __future__ import division as __division__
-from . import aperture, interferometer_seidel,interferometer_zenike
+from . import aperture, interferometer_seidel, interferometer_zenike
 from . import seidel, seidel2, zernike, test, tools, diffraction, jones, gauss
 from . import phaseunwrap, lens, asphere, mplot3d, zernike_rec
 import warnings as __warnings__
